full and simple camera models before refactoring
The full and simple camera model notebooks seem to work, but there is a problem with initialization of the sampling in nutpie in the full camera model. After a lot of debugging I decided to refactor the vonMises distribution to a normalized Gaussian. This should work well enough for our purposes. This is the lsat commit before refactoring